feat: add dashboard route, page

+++ client/src/features/invoices/pages/InvoiceListPage.jsx
@@ -26,12 +26,13 @@ import {
   useGetAllInvoicesQuery,
 } from "../invoiceApiSlice";
 import StyledDivider from "../../../components/StyledDivider";
-import { FaFileInvoiceDollar } from "react-icons/fa6";
 import { MdAssignmentAdd } from "react-icons/md";
 import Paper from "@mui/material/Paper";
-import { TiDocumentDelete } from "react-icons/ti";
+import DeleteIcon from '@mui/icons-material/Delete';
 import addBillSvg from "../../../assets/images/add_bill.svg";
 import '../../../assets/styles/customer-button.css'
+import FileOpenIcon from '@mui/icons-material/FileOpen';
+import { LiaFileInvoiceDollarSolid} from "react-icons/lia";
 
 const InvoicePage = () => {
   const navigate = useNavigate();
@@ -59,6 +60,8 @@ const InvoicePage = () => {
   };
 
   const handleDelete = async (id) => {
+    
+    if(!window.confirm('Are you sure you want to delete this invoice?')) return;
     try {
       const result = await deleteInvoice(id).unwrap();
       toast.success(result.message);
@@ -78,9 +81,11 @@ const InvoicePage = () => {
           flexDirection: "row",
           justifyContent: "center",
           alignItems: "center",
+          fontSize:"6rem",
         }}
       >
-        <Typography variant="h1"> Invoices </Typography>
+        <LiaFileInvoiceDollarSolid/>
+        <Typography pl="5px" variant="h1"> Invoices </Typography>
       </Box>
       <StyledDivider />
       <Stack
@@ -88,6 +93,7 @@ const InvoicePage = () => {
           display: "flex",
           flexDirection: "row",
           justifyContent: "space-between",
+          m:4
         }}
       >
         <Stack direction="row">
@@ -97,7 +103,7 @@ const InvoicePage = () => {
             color="primary"
             sx={{ marginTop: "3px", marginLeft: "5px" }}
           >
-            <FaFileInvoiceDollar size={40} color="action" fontSize="large" />
+            <LiaFileInvoiceDollarSolid size={40} color="action" fontSize="large" />
           </Badge>
         </Stack>
 
@@ -131,13 +137,13 @@ const InvoicePage = () => {
                 <StyledTableCell align="right">Email</StyledTableCell>
                 <StyledTableCell align="right">Status</StyledTableCell>
                 <StyledTableCell align="right">Amount</StyledTableCell>
+                <StyledTableCell align="right">Open</StyledTableCell>
                 <StyledTableCell align="right">Delete Invoice</StyledTableCell>
               </TableRow>
             </TableHead>
             <TableBody>
               {rows?.map((row) => (
                 <TableRow
-                  onClick={() => navigate(`/invoices/${row.id}`)}
                   key={row.id}
                 >
                   <TableCell style={{ width: 160 }}>
@@ -153,9 +159,13 @@ const InvoicePage = () => {
                     {row.totalAmount}
                   </TableCell>
 
+                  <TableCell  sx={{ width: 160, }}>
+                    <FileOpenIcon sx={{cursor:'pointer'}} onClick={() => navigate(`/invoices/${row.id}`)}/>
+                  </TableCell>
+
                   <TableCell style={{ width: 160 }} align="left">
                     {/* delete user account */}
-                    <TiDocumentDelete
+                    <DeleteIcon
                       sx={{ cursor: "pointer" }}
                       color="error"
                       fontSize="medium"
